==> code/02_clean_cdp_housing_unit_data.Rmd <==
---
title: "Clean census-designated place total housing unit data"
author: Shenyue Jia
output: html_notebook
---

- This notebook includes steps to ensure the names of census-designated places (CDP) match with the format used in RDS data of city level.

```{r setup, warning=FALSE}
library(tidyverse)
library(zoo)
library(foreign)
library(here)
```

```{r}
rm(list = ls())
```

- Read the table with CDP and county names and CDP housing units data
```{r}
df_cdp_cnty <- read_csv(here("data_private","CDPs_county_names_full.csv"))
df_cdp_hu <- read_csv(here("data_private","CDPs_housing_units.csv"))
```

- Inspect field names of both data frames
```{r}
head(df_cdp_cnty)
```

```{r}
head(df_cdp_hu)
```
- Join two dataframes
```{r}

df_cdp_hu <- df_cdp_hu %>%
    left_join(df_cdp_cnty,by=c("GEOID_short"="GEOID"))
```

- Drop some irrelevant fields
```{r}
df_cdp_hu <- df_cdp_hu%>%
    select(STATEFP,PLACEFP,GEOID_short,NAME,NAMELSAD,COUNTYFP,GEOID_2,NAME_2,
           NAMELSAD_2, total_housing_unit,owner_occupied,renter_occupied) %>%
    rename(GEOID_place = GEOID_short,
           name_place = NAME,
           name_place_lsad = NAMELSAD,
           GEOID_county = GEOID_2,
           name_county = NAME_2,
           name_county_lsad = NAMELSAD_2)
```

- Drop unmatched rows
```{r}
df_cdp_hu <- df_cdp_hu %>%
    filter(!is.na(GEOID_place))
```

- Write the new CDP housing unit data with county names
```{r}
write_csv(df_cdp_hu,file = 
              here("data_private",
                   "housing-units-cdp-2020-cnty-names-california.csv"))
```
