==> code/08_wui_area_in_cdp.Rmd <==
---
title: "Calculate area in WUI for CDPs in California"
author: Shenyue Jia
output: html_notebook
---

This notebook calculates the percentage of area in WUI for CDPs in California

```{r setup, warning=FALSE}
# loading libraries
library(tidyverse)
library(zoo)
library(foreign)
library(here)
library(purrr)
library(lubridate)

```

- Clean workspace
```{r}
rm(list = ls())
```


```{r}
df <- read_csv(here("data_private","wui_place_intersection_place_name.csv"))
head(df)
```

- Select some fields
```{r}
df <- df %>%
    select(Shape_Area, ALAND, NAME, NAMELSAD)
```

- Group by `NAME` and `NAMELSAD`

```{r}

df <- df %>%
    group_by(NAME,NAMELSAD) %>%
    summarise(wui_area = sum(Shape_Area),
              cdp_area = first(ALAND),
              pct_wui_area = round(wui_area/cdp_area,2)*100)
head(df)
```

- Save as a result
```{r}
write_csv(df, here("result","wui_cdp_percentage.csv"))
```


Add a new chunk by clicking the *Insert Chunk* button on the toolbar or by pressing *Cmd+Option+I*.

When you save the notebook, an HTML file containing the code and output will be saved alongside it (click the *Preview* button or press *Cmd+Shift+K* to preview the HTML file). 

The preview shows you a rendered HTML copy of the contents of the editor. Consequently, unlike *Knit*, *Preview* does not run any R code chunks. Instead, the output of the chunk when it was last run in the editor is displayed.
