==> code/14_emPOWER_data_prep.Rmd <==
---
title: "Prepare HHS emPOWER data for medically vulnerable populations"
authors: Shenyue Jia
output: html_notebook
---

## HHS emPOWER data

- HHS [emPOWER](https://empowerprogram.hhs.gov/about-empowermap.html) data tracks the change of Medicare beneficiaries from month to month since 2016
    - Data available at ZIP code, county, and state level
    - Number and percentage of durable medical equipment (DME) users are also provided
- We will work on CSV data at **county** level from 2017 to 2022 in California

```{r setup, warning=FALSE}
# loading libraries
library(tidyverse)
library(zoo)
library(foreign)
library(here)
library(purrr)
library(lubridate)
library(ggthemes)
library(hrbrthemes)
library(classInt)
library(stringr)
```

- Clean workspace
```{r}
rm(list = ls())
```

### Load data

- Bind rows of all years of emPOWER county level data
```{r}
csv_file_names <- list.files(here("data_private","emPOWER"),
                             pattern = "\\.csv$", ignore.case = TRUE)

here("data_private","emPOWER",csv_file_names) %>%
    map_dfr(read_csv) -> df_empower 

df_empower_ca <- df_empower%>%
    filter(State == 'CA')

names(df_empower)  
```

## Prepare data frame for Medicare beneficiaries
```{r}
col_names <- names(df_empower)
col_names_long <- col_names[endsWith(col_names,"Medicare_Benes")]
```

```{r}
# df_medicare_bene_wide <- df_empower[,
#                                    c(col_names[seq(1:5)],
#                                      col_names[endsWith(col_names,"Medicare_Benes")])] %>%
#     filter(State == 'CA') %>%
#     select(-FIPS_Code,-County_FIPS_Code,-County,-State_FIPS_Code,-State)

df_medicare_bene_wide <- df_empower[,
                                   c(col_names[seq(1:5)],
                                     col_names[endsWith(col_names,"Medicare_Benes")])] %>%
    filter(State == 'CA')

df_medicare_bene_long <- df_medicare_bene_wide %>%
    pivot_longer(
        cols = ends_with("Benes"),
        names_to = "month_year",
        values_to = "value"
    ) %>%
    mutate(month = substr(month_year,1,3),
           year = substr(month_year,5,8)) %>%
    select(-month_year) %>%
    drop_na()
```

- Obtain `Dec` data
```{r}
df_medicare_bene_long_dec <- df_medicare_bene_long %>%
    filter(month == 'Dec')
```

## Prepare dataframe for DME users
```{r}
col_names_long <- col_names[endsWith(col_names,"DME")]
```

```{r}
# df_medicare_bene_wide <- df_empower[,
#                                    c(col_names[seq(1:5)],
#                                      col_names[endsWith(col_names,"Medicare_Benes")])] %>%
#     filter(State == 'CA') %>%
#     select(-FIPS_Code,-County_FIPS_Code,-County,-State_FIPS_Code,-State)

df_dme_wide <- df_empower[,c(col_names[seq(1:5)],col_names[endsWith(col_names,"DME")])] %>%
    filter(State == 'CA')

df_dme_long <- df_dme_wide %>%
    pivot_longer(
        cols = ends_with("DME"),
        names_to = "month_year",
        values_to = "value"
    ) %>%
    mutate(month = substr(month_year,1,3),
           year = substr(month_year,5,8)) %>%
    select(-month_year) %>%
    drop_na()
```

- Obtain `Dec` data
```{r}
df_dme_long_dec <- df_dme_long %>%
    filter(month == 'Dec')
```

## Save to external file
```{r}
write_csv(df_dme_long, 
          here("data_private","emPOWER","dme_long_form_2017_2022.csv"))
write_csv(df_medicare_bene_long, 
          here("data_private","emPOWER","medicare_bene_long_form_2017_2022.csv"))
```
